Adding a new file for compiling in 64 bits double precision with Inter Fortran. Updating cmake files and files of the library that where in a conflicting state for working with double precision.
git-svn-id: https://forge.ill.fr/svn/crysfml@787 06508225-4b59-0410-a2a9-fdab55894d57

+++ Src/CFML_Export_Vtk.f90
@@ -194,7 +194,7 @@ Module CFML_Export_VTK
     Logical :: i_error
     !real(kind=cp), dimension(3,3) :: ort_mat
     real(kind=cp), dimension(3)   :: xx, v
-    real, dimension(:,:), allocatable :: xtemp
+    real(kind=cp), dimension(:,:), allocatable :: xtemp
 
     open(unit=pdb_id,file=trim(filename)//'.pdb', status='REPLACE',  iostat=ier)
     if (ier/=0) i_error=.true.
